完善床头灯相关功能

+++ app/src/main/java/com/ihomey/linkuphome/widget/DonutProgress.java
@@ -216,7 +216,7 @@ public class DonutProgress extends View {
         return getProgress() / (float) max * 360f;
     }
 
-    public float getProgress() {
+    public int getProgress() {
         return progress;
     }
